requirements_cloud.txt
flask>=3.0.0
anthropic>=0.25.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.11.0
eventlet>=0.35.0
